projeto-integrador/backend/requirements.txt
Django>=5.0,<6.0
psycopg2-binary>=2.9
djangorestframework>=3.15
djangorestframework-simplejwt>=5.2.2
Pillow>=9.0.0
gunicorn>=21.2.0
django-cors-headers>=4.3.1
